Reset body before reusing request
Since we are reusing a request where the Body buffer has
already been read, we need to call GetBody() to get a new
copy of the body and copy the request object. This allows
correct reading of the body in the subsequent request
decorated with the Authorization headers.

Without this change, the client sends empty body messages
on the decorated authorized requests.

This only happens when TLS verification is disabled:
http.DefaultTransport.(*http.Transport).TLSClientConfig = &tls.Config{InsecureSkipVerify: true}