regenerate client using go-swagger v0.25
Versions prior to go-swagger v0.22 (we were using v0.19) had no support
for go modules and did not as such generate the proper module paths with
the v2 included.

In this migration a few changes had to be made in the templates but they
were seemingly small.

For reference some of the changes are described in the commit
b99967394036105c3d6cbbfd604b13645f651ca2 of go-swagger.

In addition the package form3_test was merged into form3.