Implement runtime.ClientResponseStatus in our templated responses
This allows for any API response to be tested in a common way. e.g. to identify a conflict you could now do:
var clientResponseStatus runtime.ClientResponseStatus
ok := errors.As(err, &clientResponseStatus)
if ok && clientResponseStatus.IsCode(http.StatusConflict) {
...

All of our generated responses now meet this, and any status codes received that are not defined in Swagger will be *runtime.APIError, which also implements this interface.

This is a feature that is now in the default go-swagger templates, but wasn't in the Form3 templates. This is becoming a commonly used feature and adding it makes us compatible.