Remove list payments amount filter in tests
There's no guarantee that the environment being targeted by the tests will have a payment with that specific value, even a common scenario amount such as 600.00.

Looking at the git history for that test, the amount has been changed previously to get tests to pass. I don't see value in that filter being in the test, so it's removed to prevent this problem from recurring.