Fix subscriptions tests
deactivated assertion expected value updated to ptr, since the field was changed to be a *bool.

Updated record type subscribed to in tests from "PaymentSubmission" to "payment_submissions". This snake_case resource name is what's used in newer cloud-agnostic stacks. Since that is now the majority of stacks deployed, I've updated it to the new style.